+++ include/effects/Brightness.h
@@ -90,7 +90,7 @@ namespace openshot
 		/// @returns The modified openshot::Frame object
 		/// @param frame The frame object that needs the effect applied to it
 		/// @param frame_number The frame number (starting at 1) of the effect on the timeline.
-		tr1::shared_ptr<Frame> GetFrame(tr1::shared_ptr<Frame> frame, int frame_number);
+		tr1::shared_ptr<Frame> GetFrame(tr1::shared_ptr<Frame> frame, long int frame_number);
 
 		/// Get and Set JSON methods
 		string Json(); ///< Generate JSON string of this object
@@ -100,7 +100,7 @@ namespace openshot
 
 		/// Get all properties for a specific frame (perfect for a UI to display the current state
 		/// of all properties at any time)
-		string PropertiesJSON(int requested_frame);
+		string PropertiesJSON(long int requested_frame);
 	};
 
 }
