+++ src/ChunkReader.cpp
@@ -123,7 +123,7 @@ void ChunkReader::load_json()
 }
 
 // Find the location of a frame in a chunk
-ChunkLocation ChunkReader::find_chunk_frame(int requested_frame)
+ChunkLocation ChunkReader::find_chunk_frame(long int requested_frame)
 {
 	// Determine which chunk contains this frame.
 	int chunk_number = (requested_frame / chunk_size) + 1;
@@ -187,7 +187,7 @@ string ChunkReader::get_chunk_path(int chunk_number, string folder, string exten
 }
 
 // Get an openshot::Frame object for a specific frame number of this reader.
-tr1::shared_ptr<Frame> ChunkReader::GetFrame(int requested_frame) throw(ReaderClosed, ChunkNotFound)
+tr1::shared_ptr<Frame> ChunkReader::GetFrame(long int requested_frame) throw(ReaderClosed, ChunkNotFound)
 {
 	// Determine what chunk contains this frame
 	ChunkLocation location = find_chunk_frame(requested_frame);
