+++ src/effects/Mask.cpp
@@ -110,7 +110,7 @@ tr1::shared_ptr<QImage> Mask::get_grayscale_mask(tr1::shared_ptr<QImage> mask_fr
 
 // This method is required for all derived classes of EffectBase, and returns a
 // modified openshot::Frame object
-tr1::shared_ptr<Frame> Mask::GetFrame(tr1::shared_ptr<Frame> frame, int frame_number)
+tr1::shared_ptr<Frame> Mask::GetFrame(tr1::shared_ptr<Frame> frame, long int frame_number)
 {
 	// Get the mask image (from the mask reader)
 	tr1::shared_ptr<QImage> frame_image = frame->GetImage();
@@ -265,7 +265,7 @@ void Mask::SetJsonValue(Json::Value root) {
 }
 
 // Get all properties for a specific frame
-string Mask::PropertiesJSON(int requested_frame) {
+string Mask::PropertiesJSON(long int requested_frame) {
 
 	// Requested Point
 	Point requested_point(requested_frame, requested_frame);
